Move all (inactive) API endpoint specific tests to new structure
