+++ tests/legacy/ApiTest.php
@@ -868,29 +868,6 @@ class ApiTest extends FixtureTest
 	}
 
 	/**
-	 * Test the api_account_verify_credentials() function.
-	 *
-	 * @return void
-	 */
-	public function testApiAccountVerifyCredentials()
-	{
-		// self::assertArrayHasKey('user', api_account_verify_credentials('json'));
-	}
-
-	/**
-	 * Test the api_account_verify_credentials() function without an authenticated user.
-	 *
-	 * @return void
-	 */
-	public function testApiAccountVerifyCredentialsWithoutAuthenticatedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// $_SESSION['authenticated'] = false;
-		// api_account_verify_credentials('json');
-	}
-
-	/**
 	 * Test the api_statuses_mediap() function.
 	 *
 	 * @return void
@@ -1082,519 +1059,6 @@ class ApiTest extends FixtureTest
 	}
 
 	/**
-	 * Test the api_status_show() function.
-	 */
-	public function testApiStatusShowWithJson()
-	{
-		// $result = api_status_show('json', 1);
-		// self::assertStatus($result['status']);
-	}
-
-	/**
-	 * Test the api_status_show() function with an XML result.
-	 */
-	public function testApiStatusShowWithXml()
-	{
-		// $result = api_status_show('xml', 1);
-		// self::assertXml($result, 'statuses');
-	}
-
-	/**
-	 * Test the api_get_last_status() function
-	 */
-	public function testApiGetLastStatus()
-	{
-		// $item = api_get_last_status($this->selfUser['id'], $this->selfUser['id']);
-		// self::assertNotNull($item);
-	}
-
-	/**
-	 * Test the api_users_show() function.
-	 *
-	 * @return void
-	 */
-	public function testApiUsersShow()
-	{
-		/*
-		$result = api_users_show('json');
-		// We can't use assertSelfUser() here because the user object is missing some properties.
-		self::assertEquals($this->selfUser['id'], $result['user']['cid']);
-		self::assertEquals('DFRN', $result['user']['location']);
-		self::assertEquals($this->selfUser['name'], $result['user']['name']);
-		self::assertEquals($this->selfUser['nick'], $result['user']['screen_name']);
-		self::assertTrue($result['user']['verified']);
-		*/
-	}
-
-	/**
-	 * Test the api_users_show() function with an XML result.
-	 *
-	 * @return void
-	 */
-	public function testApiUsersShowWithXml()
-	{
-		// $result = api_users_show('xml');
-		// self::assertXml($result, 'statuses');
-	}
-
-	/**
-	 * Test the api_users_search() function.
-	 *
-	 * @return void
-	 */
-	public function testApiUsersSearch()
-	{
-		// $_GET['q'] = 'othercontact';
-		// $result    = api_users_search('json');
-		// self::assertOtherUser($result['users'][0]);
-	}
-
-	/**
-	 * Test the api_users_search() function with an XML result.
-	 *
-	 * @return void
-	 */
-	public function testApiUsersSearchWithXml()
-	{
-		// $_GET['q'] = 'othercontact';
-		// $result    = api_users_search('xml');
-		// self::assertXml($result, 'users');
-	}
-
-	/**
-	 * Test the api_users_search() function without a GET q parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiUsersSearchWithoutQuery()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\BadRequestException::class);
-		// api_users_search('json');
-	}
-
-	/**
-	 * Test the api_users_lookup() function.
-	 *
-	 * @return void
-	 */
-	public function testApiUsersLookup()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\NotFoundException::class);
-		// api_users_lookup('json');
-	}
-
-	/**
-	 * Test the api_users_lookup() function with an user ID.
-	 *
-	 * @return void
-	 */
-	public function testApiUsersLookupWithUserId()
-	{
-		// $_REQUEST['user_id'] = $this->otherUser['id'];
-		// $result              = api_users_lookup('json');
-		// self::assertOtherUser($result['users'][0]);
-	}
-
-	/**
-	 * Test the api_search() function.
-	 *
-	 * @return void
-	 */
-	public function testApiSearch()
-	{
-		/*
-		$_REQUEST['q']      = 'reply';
-		$_REQUEST['max_id'] = 10;
-		$result             = api_search('json');
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-			self::assertStringContainsStringIgnoringCase('reply', $status['text'], '', true);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_search() function a count parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiSearchWithCount()
-	{
-		/*
-		$_REQUEST['q']     = 'reply';
-		$_REQUEST['count'] = 20;
-		$result            = api_search('json');
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-			self::assertStringContainsStringIgnoringCase('reply', $status['text'], '', true);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_search() function with an rpp parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiSearchWithRpp()
-	{
-		/*
-		$_REQUEST['q']   = 'reply';
-		$_REQUEST['rpp'] = 20;
-		$result          = api_search('json');
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-			self::assertStringContainsStringIgnoringCase('reply', $status['text'], '', true);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_search() function with an q parameter contains hashtag.
-	 * @doesNotPerformAssertions
-	 */
-	public function testApiSearchWithHashtag()
-	{
-		/*
-		$_REQUEST['q'] = '%23friendica';
-		$result        = api_search('json');
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-			self::assertStringContainsStringIgnoringCase('#friendica', $status['text'], '', true);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_search() function with an exclude_replies parameter.
-	 * @doesNotPerformAssertions
-	 */
-	public function testApiSearchWithExcludeReplies()
-	{
-		/*
-		$_REQUEST['max_id']          = 10;
-		$_REQUEST['exclude_replies'] = true;
-		$_REQUEST['q']               = 'friendica';
-		$result                      = api_search('json');
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_search() function without an authenticated user.
-	 *
-	 * @return void
-	 */
-	public function testApiSearchWithUnallowedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// api_search('json');
-	}
-
-	/**
-	 * Test the api_search() function without any GET query parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiSearchWithoutQuery()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\BadRequestException::class);
-		// api_search('json');
-	}
-
-	/**
-	 * Test the api_statuses_home_timeline() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesHomeTimeline()
-	{
-		/*
-		$_REQUEST['max_id']          = 10;
-		$_REQUEST['exclude_replies'] = true;
-		$_REQUEST['conversation_id'] = 1;
-		$result                      = api_statuses_home_timeline('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_home_timeline() function with a negative page parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesHomeTimelineWithNegativePage()
-	{
-		/*
-		$_REQUEST['page'] = -2;
-		$result           = api_statuses_home_timeline('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_home_timeline() with an unallowed user.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesHomeTimelineWithUnallowedUser()
-	{
-		/*
-		$this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		BasicAuth::setCurrentUserID();
-		api_statuses_home_timeline('json');
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_home_timeline() function with an RSS result.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesHomeTimelineWithRss()
-	{
-		// $result = api_statuses_home_timeline('rss');
-		// self::assertXml($result, 'statuses');
-	}
-
-	/**
-	 * Test the api_statuses_public_timeline() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesPublicTimeline()
-	{
-		/*
-		$_REQUEST['max_id']          = 10;
-		$_REQUEST['conversation_id'] = 1;
-		$result                      = api_statuses_public_timeline('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_public_timeline() function with the exclude_replies parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesPublicTimelineWithExcludeReplies()
-	{
-		/*
-		$_REQUEST['max_id']          = 10;
-		$_REQUEST['exclude_replies'] = true;
-		$result                      = api_statuses_public_timeline('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_public_timeline() function with a negative page parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesPublicTimelineWithNegativePage()
-	{
-		/*
-		$_REQUEST['page'] = -2;
-		$result           = api_statuses_public_timeline('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_public_timeline() function with an unallowed user.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesPublicTimelineWithUnallowedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// api_statuses_public_timeline('json');
-	}
-
-	/**
-	 * Test the api_statuses_public_timeline() function with an RSS result.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesPublicTimelineWithRss()
-	{
-		// $result = api_statuses_public_timeline('rss');
-		// self::assertXml($result, 'statuses');
-	}
-
-	/**
-	 * Test the api_statuses_networkpublic_timeline() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesNetworkpublicTimeline()
-	{
-		/*
-		$_REQUEST['max_id'] = 10;
-		$result             = api_statuses_networkpublic_timeline('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_networkpublic_timeline() function with a negative page parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesNetworkpublicTimelineWithNegativePage()
-	{
-		/*
-		$_REQUEST['page'] = -2;
-		$result           = api_statuses_networkpublic_timeline('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_networkpublic_timeline() function with an unallowed user.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesNetworkpublicTimelineWithUnallowedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// api_statuses_networkpublic_timeline('json');
-	}
-
-	/**
-	 * Test the api_statuses_networkpublic_timeline() function with an RSS result.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesNetworkpublicTimelineWithRss()
-	{
-		// $result = api_statuses_networkpublic_timeline('rss');
-		// self::assertXml($result, 'statuses');
-	}
-
-	/**
-	 * Test the api_statuses_show() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesShow()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\BadRequestException::class);
-		// api_statuses_show('json');
-	}
-
-	/**
-	 * Test the api_statuses_show() function with an ID.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesShowWithId()
-	{
-		// DI::args()->setArgv(['', '', '', 1]);
-		// $result = api_statuses_show('json');
-		// self::assertStatus($result['status']);
-	}
-
-	/**
-	 * Test the api_statuses_show() function with the conversation parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesShowWithConversation()
-	{
-		/*
-		DI::args()->setArgv(['', '', '', 1]);
-		$_REQUEST['conversation'] = 1;
-		$result                   = api_statuses_show('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_show() function with an unallowed user.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesShowWithUnallowedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// api_statuses_show('json');
-	}
-
-	/**
-	 * Test the api_conversation_show() function.
-	 *
-	 * @return void
-	 */
-	public function testApiConversationShow()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\BadRequestException::class);
-		// api_conversation_show('json');
-	}
-
-	/**
-	 * Test the api_conversation_show() function with an ID.
-	 *
-	 * @return void
-	 */
-	public function testApiConversationShowWithId()
-	{
-		/*
-		DI::args()->setArgv(['', '', '', 1]);
-		$_REQUEST['max_id'] = 10;
-		$_REQUEST['page']   = -2;
-		$result             = api_conversation_show('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_conversation_show() function with an unallowed user.
-	 *
-	 * @return void
-	 */
-	public function testApiConversationShowWithUnallowedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// api_conversation_show('json');
-	}
-
-	/**
 	 * Test the api_statuses_repeat() function.
 	 *
 	 * @return void
@@ -1633,156 +1097,6 @@ class ApiTest extends FixtureTest
 		DI::args()->setArgv(['', '', '', 5]);
 		$result = api_statuses_repeat('json');
 		self::assertStatus($result['status']);
-	}
-
-	/**
-	 * Test the api_statuses_destroy() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesDestroy()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\BadRequestException::class);
-		// api_statuses_destroy('json');
-	}
-
-	/**
-	 * Test the api_statuses_destroy() function without an authenticated user.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesDestroyWithoutAuthenticatedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// $_SESSION['authenticated'] = false;
-		// api_statuses_destroy('json');
-	}
-
-	/**
-	 * Test the api_statuses_destroy() function with an ID.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesDestroyWithId()
-	{
-		// DI::args()->setArgv(['', '', '', 1]);
-		// $result = api_statuses_destroy('json');
-		// self::assertStatus($result['status']);
-	}
-
-	/**
-	 * Test the api_statuses_mentions() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesMentions()
-	{
-		/*
-		$this->app->setLoggedInUserNickname($this->selfUser['nick']);
-		$_REQUEST['max_id'] = 10;
-		$result             = api_statuses_mentions('json');
-		self::assertEmpty($result['status']);
-		// We should test with mentions in the database.
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_mentions() function with a negative page parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesMentionsWithNegativePage()
-	{
-		// $_REQUEST['page'] = -2;
-		// $result           = api_statuses_mentions('json');
-		// self::assertEmpty($result['status']);
-	}
-
-	/**
-	 * Test the api_statuses_mentions() function with an unallowed user.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesMentionsWithUnallowedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// api_statuses_mentions('json');
-	}
-
-	/**
-	 * Test the api_statuses_mentions() function with an RSS result.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesMentionsWithRss()
-	{
-		// $result = api_statuses_mentions('rss');
-		// self::assertXml($result, 'statuses');
-	}
-
-	/**
-	 * Test the api_statuses_user_timeline() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesUserTimeline()
-	{
-	/*
-		$_REQUEST['user_id']         = 42;
-		$_REQUEST['max_id']          = 10;
-		$_REQUEST['exclude_replies'] = true;
-		$_REQUEST['conversation_id'] = 7;
-
-		$result = api_statuses_user_timeline('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_user_timeline() function with a negative page parameter.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesUserTimelineWithNegativePage()
-	{
-		/*
-		$_REQUEST['user_id'] = 42;
-		$_REQUEST['page']    = -2;
-
-		$result = api_statuses_user_timeline('json');
-		self::assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_statuses_user_timeline() function with an RSS result.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesUserTimelineWithRss()
-	{
-		// $result = api_statuses_user_timeline('rss');
-		// self::assertXml($result, 'statuses');
-	}
-
-	/**
-	 * Test the api_statuses_user_timeline() function with an unallowed user.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesUserTimelineWithUnallowedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// api_statuses_user_timeline('json');
 	}
 
 	/**
@@ -1875,45 +1189,7 @@ class ApiTest extends FixtureTest
 		api_favorites_create_destroy('json');
 	}
 
-	/**
-	 * Test the api_favorites() function.
-	 *
-	 * @return void
-	 */
-	public function testApiFavorites()
-	{
-		/*
-		$_REQUEST['page']   = -1;
-		$_REQUEST['max_id'] = 10;
-		$result             = api_favorites('json');
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
 
-	/**
-	 * Test the api_favorites() function with an RSS result.
-	 *
-	 * @return void
-	 */
-	public function testApiFavoritesWithRss()
-	{
-		// $result = api_favorites('rss');
-		// self::assertXml($result, 'statuses');
-	}
-
-	/**
-	 * Test the api_favorites() function with an unallowed user.
-	 *
-	 * @return void
-	 */
-	public function testApiFavoritesWithUnallowedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// api_favorites('json');
-	}
 
 	/**
 	 * Test the api_format_messages() function.
@@ -2251,92 +1527,6 @@ class ApiTest extends FixtureTest
 	}
 
 	/**
-	 * Test the api_lists_statuses() function.
-	 *
-	 * @return void
-	 */
-	public function testApiListsStatuses()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\BadRequestException::class);
-		// api_lists_statuses('json');
-	}
-
-	/**
-	 * Test the api_lists_statuses() function with a list ID.
-	 * @doesNotPerformAssertions
-	 */
-	public function testApiListsStatusesWithListId()
-	{
-		/*
-		$_REQUEST['list_id'] = 1;
-		$_REQUEST['page']    = -1;
-		$_REQUEST['max_id']  = 10;
-		$result              = api_lists_statuses('json');
-		foreach ($result['status'] as $status) {
-			self::assertStatus($status);
-		}
-		*/
-	}
-
-	/**
-	 * Test the api_lists_statuses() function with a list ID and a RSS result.
-	 *
-	 * @return void
-	 */
-	public function testApiListsStatusesWithListIdAndRss()
-	{
-		// $_REQUEST['list_id'] = 1;
-		// $result              = api_lists_statuses('rss');
-		// self::assertXml($result, 'statuses');
-	}
-
-	/**
-	 * Test the api_lists_statuses() function with an unallowed user.
-	 *
-	 * @return void
-	 */
-	public function testApiListsStatusesWithUnallowedUser()
-	{
-		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
-		// BasicAuth::setCurrentUserID();
-		// api_lists_statuses('json');
-	}
-
-	/**
-	 * Test the api_statuses_f() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesFWithFriends()
-	{
-		// $_GET['page'] = -1;
-		// $result       = api_statuses_f('friends');
-		// self::assertArrayHasKey('user', $result);
-	}
-
-	/**
-	 * Test the api_statuses_f() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesFWithFollowers()
-	{
-		// $result = api_statuses_f('followers');
-		// self::assertArrayHasKey('user', $result);
-	}
-
-	/**
-	 * Test the api_statuses_f() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesFWithBlocks()
-	{
-		// $result = api_statuses_f('blocks');
-		// self::assertArrayHasKey('user', $result);
-	}
-
-	/**
 	 * Test the api_statuses_f() function.
 	 *
 	 * @return void
@@ -2347,126 +1537,6 @@ class ApiTest extends FixtureTest
 		// self::assertArrayHasKey('user', $result);
 	}
 
-	/**
-	 * Test the api_statuses_f() function an undefined cursor GET variable.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesFWithUndefinedCursor()
-	{
-		// $_GET['cursor'] = 'undefined';
-		// self::assertFalse(api_statuses_f('friends'));
-	}
-
-	/**
-	 * Test the api_statuses_friends() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesFriends()
-	{
-		// $result = api_statuses_friends('json');
-		// self::assertArrayHasKey('user', $result);
-	}
-
-	/**
-	 * Test the api_statuses_friends() function an undefined cursor GET variable.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesFriendsWithUndefinedCursor()
-	{
-		// $_GET['cursor'] = 'undefined';
-		// self::assertFalse(api_statuses_friends('json'));
-	}
-
-	/**
-	 * Test the api_statuses_followers() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesFollowers()
-	{
-		// $result = api_statuses_followers('json');
-		// self::assertArrayHasKey('user', $result);
-	}
-
-	/**
-	 * Test the api_statuses_followers() function an undefined cursor GET variable.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusesFollowersWithUndefinedCursor()
-	{
-		// $_GET['cursor'] = 'undefined';
-		// self::assertFalse(api_statuses_followers('json'));
-	}
-
-	/**
-	 * Test the api_blocks_list() function.
-	 *
-	 * @return void
-	 */
-	public function testApiBlocksList()
-	{
-		// $result = api_blocks_list('json');
-		// self::assertArrayHasKey('user', $result);
-	}
-
-	/**
-	 * Test the api_blocks_list() function an undefined cursor GET variable.
-	 *
-	 * @return void
-	 */
-	public function testApiBlocksListWithUndefinedCursor()
-	{
-		// $_GET['cursor'] = 'undefined';
-		// self::assertFalse(api_blocks_list('json'));
-	}
-
-	/**
-	 * Test the api_friendships_incoming() function.
-	 *
-	 * @return void
-	 */
-	public function testApiFriendshipsIncoming()
-	{
-		// $result = api_friendships_incoming('json');
-		// self::assertArrayHasKey('id', $result);
-	}
-
-	/**
-	 * Test the api_friendships_incoming() function an undefined cursor GET variable.
-	 *
-	 * @return void
-	 */
-	public function testApiFriendshipsIncomingWithUndefinedCursor()
-	{
-		// $_GET['cursor'] = 'undefined';
-		// self::assertFalse(api_friendships_incoming('json'));
-	}
-
-	/**
-	 * Test the api_statusnet_config() function.
-	 *
-	 * @return void
-	 */
-	public function testApiStatusnetConfig()
-	{
-		/*
-		$result = api_statusnet_config('json');
-		self::assertEquals('localhost', $result['config']['site']['server']);
-		self::assertEquals('default', $result['config']['site']['theme']);
-		self::assertEquals(DI::baseUrl() . '/images/friendica-64.png', $result['config']['site']['logo']);
-		self::assertTrue($result['config']['site']['fancy']);
-		self::assertEquals('en', $result['config']['site']['language']);
-		self::assertEquals('UTC', $result['config']['site']['timezone']);
-		self::assertEquals(200000, $result['config']['site']['textlimit']);
-		self::assertEquals('false', $result['config']['site']['private']);
-		self::assertEquals('false', $result['config']['site']['ssl']);
-		self::assertEquals(30, $result['config']['site']['shorturllength']);
-		*/
-	}
 
 	/**
 	 * Test the api_direct_messages_new() function.
@@ -2950,27 +2020,6 @@ class ApiTest extends FixtureTest
 	{
 		$this->expectException(\Friendica\Network\HTTPException\BadRequestException::class);
 		$this->markTestIncomplete();
-	}
-
-
-	/**
-	 * Test the api_account_update_profile() function.
-	 *
-	 * @return void
-	 */
-	public function testApiAccountUpdateProfile()
-	{
-		/*
-		$_POST['name']        = 'new_name';
-		$_POST['description'] = 'new_description';
-		$result               = api_account_update_profile('json');
-		// We can't use assertSelfUser() here because the user object is missing some properties.
-		self::assertEquals($this->selfUser['id'], $result['user']['cid']);
-		self::assertEquals('DFRN', $result['user']['location']);
-		self::assertEquals($this->selfUser['nick'], $result['user']['screen_name']);
-		self::assertEquals('new_name', $result['user']['name']);
-		self::assertEquals('new_description', $result['user']['description']);
-		*/
 	}
 
 	/**
