+++ tests/src/Module/Api/GnuSocial/GnuSocial/ConfigTest.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Friendica\Test\src\Module\Api\GnuSocial\GnuSocial;
+
+use Friendica\Test\src\Module\Api\ApiTest;
+
+class ConfigTest extends ApiTest
+{
+	/**
+	 * Test the api_statusnet_config() function.
+	 *
+	 * @return void
+	 */
+	public function testApiStatusnetConfig()
+	{
+		/*
+		$result = api_statusnet_config('json');
+		self::assertEquals('localhost', $result['config']['site']['server']);
+		self::assertEquals('default', $result['config']['site']['theme']);
+		self::assertEquals(DI::baseUrl() . '/images/friendica-64.png', $result['config']['site']['logo']);
+		self::assertTrue($result['config']['site']['fancy']);
+		self::assertEquals('en', $result['config']['site']['language']);
+		self::assertEquals('UTC', $result['config']['site']['timezone']);
+		self::assertEquals(200000, $result['config']['site']['textlimit']);
+		self::assertEquals('false', $result['config']['site']['private']);
+		self::assertEquals('false', $result['config']['site']['ssl']);
+		self::assertEquals(30, $result['config']['site']['shorturllength']);
+		*/
+	}
+}
