+++ tests/src/Module/Api/Twitter/Statuses/NetworkPublicTimelineTest.php
@@ -1,0 +1,65 @@
+<?php
+
+namespace Friendica\Test\src\Module\Api\Twitter\Statuses;
+
+use Friendica\Test\src\Module\Api\ApiTest;
+
+class NetworkPublicTimelineTest extends ApiTest
+{
+	/**
+	 * Test the api_statuses_networkpublic_timeline() function.
+	 *
+	 * @return void
+	 */
+	public function testApiStatusesNetworkpublicTimeline()
+	{
+		/*
+		$_REQUEST['max_id'] = 10;
+		$result             = api_statuses_networkpublic_timeline('json');
+		self::assertNotEmpty($result['status']);
+		foreach ($result['status'] as $status) {
+			self::assertStatus($status);
+		}
+		*/
+	}
+
+	/**
+	 * Test the api_statuses_networkpublic_timeline() function with a negative page parameter.
+	 *
+	 * @return void
+	 */
+	public function testApiStatusesNetworkpublicTimelineWithNegativePage()
+	{
+		/*
+		$_REQUEST['page'] = -2;
+		$result           = api_statuses_networkpublic_timeline('json');
+		self::assertNotEmpty($result['status']);
+		foreach ($result['status'] as $status) {
+			self::assertStatus($status);
+		}
+		*/
+	}
+
+	/**
+	 * Test the api_statuses_networkpublic_timeline() function with an unallowed user.
+	 *
+	 * @return void
+	 */
+	public function testApiStatusesNetworkpublicTimelineWithUnallowedUser()
+	{
+		// $this->expectException(\Friendica\Network\HTTPException\UnauthorizedException::class);
+		// BasicAuth::setCurrentUserID();
+		// api_statuses_networkpublic_timeline('json');
+	}
+
+	/**
+	 * Test the api_statuses_networkpublic_timeline() function with an RSS result.
+	 *
+	 * @return void
+	 */
+	public function testApiStatusesNetworkpublicTimelineWithRss()
+	{
+		// $result = api_statuses_networkpublic_timeline('rss');
+		// self::assertXml($result, 'statuses');
+	}
+}
